requirements.txt
functions-framework==3.4.0
python-dotenv==1.0.0
google-api-python-client==2.95.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-auth==2.17.1
google-cloud-dns==0.34.1
